resolve conflicts, move diffsinger phonemizers into a subfolder
