+++ OpenUtau.Core/DiffSinger/Phonemizers/DiffSingerJyutpingPhonemizer.cs
@@ -6,6 +6,7 @@ using System.Linq;
 namespace OpenUtau.Core.DiffSinger {
     [Phonemizer("DiffSinger Jyutping Phonemizer", "DIFFS ZH-YUE", language: "ZH-YUE")]
     public class DiffSingerJyutpingPhonemizer : DiffSingerBasePhonemizer {
+        protected override string GetDictionaryName() => "dsdict-zh-yue.yaml";
         protected override string[] Romanize(IEnumerable<string> lyrics) {
             return ZhG2p.CantoneseInstance.Convert(lyrics.ToList(), false, true).Split(" ");
         }
